requirements.txt
aioflask==0.4.0
anyio==4.2.0
asgiref==3.7.2
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Cython==3.0.7
Flask==3.0.0
greenlet==3.0.3
greenletio==0.11.0
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
python-telegram-bot==20.0
requests==2.31.0
rfc3986==1.5.0
sniffio==1.3.0
urllib3==2.1.0
uvicorn==0.25.0
Werkzeug==3.0.1
pytz==2022.1